Restructure testcase in model
Restructure testcase to fit the directory
structure of the source code of model.
