+++ src/test/java/seedu/ddd/model/contact/common/AddressTest.java
@@ -1,12 +1,10 @@
-package seedu.ddd.model.contact;
+package seedu.ddd.model.contact.common;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.ddd.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
-
-import seedu.ddd.model.contact.common.Address;
 
 public class AddressTest {
