+++ src/test/java/seedu/ddd/model/contact/common/UniqueContactListTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.model.contact;
+package seedu.ddd.model.contact.common;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -15,8 +15,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.ddd.model.contact.common.Contact;
-import seedu.ddd.model.contact.common.UniqueContactList;
 import seedu.ddd.model.contact.exceptions.ContactNotFoundException;
 import seedu.ddd.model.contact.exceptions.DuplicateContactException;
 import seedu.ddd.testutil.ClientBuilder;
